# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PrintPrice Pro API
  description: |
    Public API for the PrintPrice Platform. Enables submission, tracking and 
    integration of PDF preflight and autofix jobs into third-party workflows.
    
    **Authentication:** All API requests must include an API key as a Bearer token:
    `Authorization: Bearer ppk_live_xxxxx`
    
    **Base URL:** `https://preflight.printprice.pro/api/v2`
  version: "1.0"
  contact:
    name: PrintPrice Support
    url: https://preflight.printprice.pro

servers:
  - url: https://preflight.printprice.pro/api/v2
    description: Production
  - url: http://localhost:8080/api/v2
    description: Development

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

  schemas:
    Job:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
          description: Unique identifier for the job.
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELED, CANCEL_REQUESTED]
        progress:
          type: integer
          description: Completion progress (0-100).
        type:
          type: string
          enum: [PREFLIGHT, AUTOFIX]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metrics:
          $ref: '#/components/schemas/JobMetrics'
        links:
          type: object
          properties:
            download_url:
              type: string
            report_url:
              type: string

    JobMetrics:
      type: object
      description: Available only when status is SUCCEEDED.
      properties:
        risk_score_before:
          type: integer
          description: Risk score of the original document (0-100).
        risk_score_after:
          type: integer
          description: Risk score after autofix (0-100).
        hours_saved:
          type: number
          format: float
          description: Estimated hours of prepress labor saved.
        value_generated:
          type: number
          format: float
          description: Estimated USD value generated by the autofix.

    Error:
      type: object
      properties:
        error:
          type: string

security:
  - BearerAuth: []

paths:
  /jobs:
    post:
      summary: Submit a PDF for Preflight + AutoFix
      operationId: createJob
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The PDF file to analyze.
                policy:
                  type: string
                  description: "Policy slug (e.g. OFFSET_CMYK_STRICT). Defaults to OFFSET_CMYK_STRICT."
      responses:
        "202":
          description: Job accepted and queued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: QUEUED
                  created_at:
                    type: string
                    format: date-time
                  links:
                    type: object
                    properties:
                      self:
                        type: string
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Invalid or missing API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List recent jobs for the authenticated tenant
      operationId: listJobs
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: A list of jobs.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
        "401":
          description: Unauthorized.

  /jobs/{id}:
    get:
      summary: Get detailed status of a specific job
      operationId: getJobById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Job detail with status and ROI metrics if completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        "401":
          description: Unauthorized.
        "404":
          description: Job not found.